Create Abstraction
Abstration means hiding implementations and also showing functionalities to be accessed for user who are going to get services from our creations. In this case, I have created a Car interface that has been the super type of "NormalCar" and "SportCar" classes which have been declared as abstract. I decided to keep those as abstract why becouse we have to use unique implementation for both of them. That means Sport cars have unique features which are not in normal cars such as Turbo powers,Nitrous gas and etc. After that "Caruser" concreate class was created to create cars. The CarUsers know functionalities but they don't know how implementation is.